pyproject.toml
[project]
name = "task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.4",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-openai>=1.1.11",
    "langgraph>=1.0.10",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
